Graph interface/design changed
Now we have the wrapper class GameGraph wich maintain an graph object of
the graph toolkit of your choice. So GameGraph objects can passed as
arguments even if the graph toolkit changes. If you want's to program to the
graph interface, call getGraph() on your GameGraph object.

For reasons I chose this design, see the documentation.