PlayerListener mrXSet method renamed
mrXSet is no good name because of the word Set. It never meant the
java.util.Set but the past of 'to set'. I think the reaseon for mrXSet was
that it can also set to null. Otherwise all triggers for ...Changed
listener methods are typical setters.
Therefore I rename mrXSet to mrXChanged.